Move test data into separate file
